started dividing Server class
