+++ include/ServerSocket.hpp
@@ -1,0 +1,48 @@
+#pragma once
+
+#include <sys/socket.h>
+#include <netinet/in.h>
+#include <unistd.h>
+#include <fcntl.h>
+#include <poll.h>
+
+#include <iostream>
+#include <vector>
+#include <string>
+#include <cstring>
+#include <csignal>
+
+
+//#include "Client.hpp"
+//#include "Request.hpp"
+//#include "Response.hpp"
+
+//const int MAX_CLIENT = 100;
+
+class ServerScoket { // rename to ServerScoketSocket maybe
+	private:
+		int fd;
+		struct sockaddr_in address;
+		int port;
+		bool isSetup;
+		//std::vector<Client> clients;
+		std::vector<pollfd> poll_fds;
+		//int client_count;
+		
+		bool createSocket();
+		bool configureSocket();
+		bool bindSocket(int port);
+		bool listenMode();
+		
+		//void handleNewConnection();
+		//void handleClientRequest(size_t index);
+		//bool sendResponse(size_t index);
+		//void cleanup();
+		//static void signalHandler(int signum); // static to be compatible with signal()
+
+	public:
+		ServerScoket();
+		//~ServerScoket();
+
+		bool setupServerScoket(int port);
+};
